Begin early attempts at writing the process_client_logs script.
We are currently running into an issue with cat * where the directory we are accepting as a command line argument
seems to not actually contain the files we need to sort through to get the log data. The files instead seem to be
located one or two subdirectories below that.

We've also got a first draft of regex for our awk command that will extract the relevant data, although we are
considering a potentially simpler solution that just plucks the specific columns of data we want when certain
conditions are met.