Get the relevant text from the log files.
Turns out the issue from before was that we were trying to pipe the results of our cd command into exit... which blew things up.

We think we have all the lines that we want from the log files, but we need to get rid of the minutes and seconds on the time given in the log files.
Our first attempt to solve this causes the program to hang up when checking whether the tests executed successfully. (syntax error?)

We've also cleaned up a bit of the unnecessary bits of code (like touch failed_login_data.txt) and changed the solution structure so we first cat all the log data in a
temporary text file, that we then scan with awk to pick up the relevant columns of information. (We'll need to remember to delete this .txt file when we are finished)