Complete the create_hours_dist.sh script.
I did this all in one shot because it is pretty much the same script as create_username_dist.sh, save
for one or two minor changes in which columns awk prints out.

I've also manually tested the graphs by wrapping them in the summary header and footer and they appear
to be working properly.